solution/pyproject.toml
[project]
name = "solution"
version = "0.1.0"
description = "A simple adventure maze game."
readme = "README.md"
authors = [
    { name = "Jettro Coenradie", email = "jettro@coenradie.com" },
    { name = "Joey Visbeen", email = "joey.visbeen@weareyuma.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "fastapi-mcp>=0.4.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "strands-agents-tools>=0.2.21",
    "strands-agents[openai]>=1.28.0",
    "streamlit>=1.54.0",
    "uvicorn>=0.41.0",
]

[project.scripts]
solution = "solution:main"

[build-system]
requires = ["uv_build>=0.10.7,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "pytest>=9.0.2",
]
